feat(bufferline): customize buffer close behavior
In a previous request (issue #234, PR #270), it was asked that
barbar.nvim mimic the behavior of a browser when closing buffers: choose
the buffer to the right of the current buffer.

* NOTE: I tested firefox to double check the behavior matched, and found
        that the current behavior _does_ work as intended.

However, we have had reports that this behavior can be irritating, and
it has been asked that we provide an option for it to be restored to its
original method of function.